requirements.txt
requests==2.32.3
feedparser==6.0.11
psycopg2-binary==2.9.9
vaderSentiment==3.3.2
spacy==3.7.5
fastapi==0.115.0
uvicorn==0.30.6
apscheduler==3.10.4
pandas==2.2.2
python-dotenv==1.0.1
